# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "THE Question Portal",
        "description": "Create a Project using laravel and reactJS",
        "termsOfService": "",
        "contact": {
            "email": "sumersingh1997.ssh@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/submit-question": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Submit Question",
                "operationId": "storeSubmitQuestion",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "question_id": {
                                        "type": "number"
                                    },
                                    "question_answer": {
                                        "type": "text"
                                    },
                                    "question_used_hints": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "406": {
                        "description": "not acceptable",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_security": []
                    }
                ]
            }
        },
        "/api/view-users": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "View Users",
                "operationId": "showUsers",
                "responses": {
                    "201": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "406": {
                        "description": "not acceptable",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_security": []
                    }
                ]
            }
        },
        "/api/view-user-questions/{id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "View Users",
                "operationId": "showUserQuestions",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User Id",
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "406": {
                        "description": "not acceptable",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_security": []
                    }
                ]
            }
        },
        "/api/get-today-question/": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get Today Question",
                "operationId": "showQuestionForTest",
                "responses": {
                    "201": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "406": {
                        "description": "not acceptable",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_security": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearer_security": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer < token >)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        []
    ]
}